pyproject.toml
[tool.poetry]
package-mode = false
name = "botagente"
version = "0.1.0"
description = "A sample project"
authors = ["John J Meza <johnometalmanldu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.28.0"
pyperclip = "^1.8.2"  # Cross-platform clipboard support
pillow = "^10.0.0"    # Cross-platform image handling
pyautogui = "^0.9.54" # Cross-platform GUI automation

# Platform-specific dependencies
pywin32 = {version = "^306", markers = "sys_platform == 'win32'"}  # Windows only
python3-xlib = {version = "^0.15", markers = "sys_platform == 'linux'"}  # Linux only
pyobjc-core = {version = "^9.2", markers = "sys_platform == 'darwin'"}  # macOS only
pyobjc-framework-Quartz = {version = "^9.2", markers = "sys_platform == 'darwin'"}  # macOS clipboard support

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
